% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractHood.R
\name{extract_hood}
\alias{extract_hood}
\title{Spiec-Easi function to extract neighborhood surrounding nodes of interest.
Input is Spiec-Easi network and node or nodes of interest as vector such as c(1:5)
This code was generalized from the code used to make figure 4 in Tipton, Müller, et. al.
written by Laura Tipton; 1/5/18}
\usage{
extract_hood(selntwrk, node)
}
\description{
Spiec-Easi function to extract neighborhood surrounding nodes of interest.
Input is Spiec-Easi network and node or nodes of interest as vector such as c(1:5)
This code was generalized from the code used to make figure 4 in Tipton, Müller, et. al.
written by Laura Tipton; 1/5/18
}
